@@ -1,48 +1,54 @@
 {
   "name": "gatsby-kodi-tv",
-  "private": true,
-  "description": "The Kodi web site",
-  "version": "0.1.0",
+  "description": "Gatsby website for kodi.tv",
+  "version": "1.0.0",
   "author": "Kyle Johnson <pkscout@kodi.tv>",
   "dependencies": {
-    "@patternfly/react-core": "^4.101.3",
+    "@headlessui/react": "^0.3.1",
+    "@mdx-js/mdx": "^1.6.22",
+    "@mdx-js/react": "^1.6.22",
+    "@stripe/react-stripe-js": "^1.4.0",
     "@stripe/stripe-js": "^1.13.1",
-    "aws-sdk": "^2.860.0",
-    "axios": "^0.21.1",
-    "eslint": "^7.21.0",
+    "@tailwindcss/forms": "^0.2.1",
+    "@tailwindcss/typography": "^0.4.0",
+    "@typescript-eslint/eslint-plugin": "^4.17.0",
+    "@typescript-eslint/parser": "^4.17.0",
+    "autoprefixer": "^10.2.5",
+    "aws-sdk": "^2.863.0",
+    "eslint": "^7.22.0",
     "gatsby": "^3.0.4",
     "gatsby-awesome-pagination": "^0.3.6",
     "gatsby-image": "^3.0.0",
     "gatsby-plugin-feed": "^3.0.0",
+    "gatsby-plugin-mdx": "^2.0.0",
     "gatsby-plugin-netlify-cms": "^5.0.0",
     "gatsby-plugin-offline": "^4.0.0",
+    "gatsby-plugin-postcss": "^4.0.0",
     "gatsby-plugin-react-helmet": "^4.0.0",
-    "gatsby-plugin-typescript": "^3.0.0",
+    "gatsby-plugin-root-import": "^2.0.6",
+    "gatsby-plugin-sass": "^4.0.2",
     "gatsby-source-filesystem": "^3.0.0",
-    "gatsby-theme-patternfly": "git+https://github.com/pkscout/gatsby-theme-patternfly.git",
     "gatsby-transformer-remark": "^3.0.0",
     "gatsby-transformer-yaml": "^3.0.0",
+    "heroicons-react": "^1.3.0",
     "js-yaml": "^4.0.0",
-    "netlify-cms-app": "^2.14.26",
+    "netlify-cms-app": "^2.14.29",
     "node-fetch": "^2.6.1",
+    "node-sass": "^5.0.0",
+    "postcss": "^8.2.8",
+    "prettier": "^2.2.1",
     "prop-types": "^15.7.2",
     "react": "^17.0.1",
-    "react-awesome-slider": "^4.1.0",
     "react-dom": "^17.0.1",
     "react-helmet": "^6.1.0",
     "react-markdown": "^5.0.3",
-    "react-social-icons": "^4.1.0",
+    "sass": "^1.32.8",
     "slugify": "^1.4.7",
+    "swiper": "^6.5.0",
+    "tailwindcss": "^2.0.3",
+    "typescript": "^4.2.3",
     "util": "^0.12.3",
     "xml-parser": "^1.2.1"
-  },
-  "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^4.17.0",
-    "@typescript-eslint/parser": "^4.17.0",
-    "eslint": "^7.21.0",
-    "gatsby-plugin-root-import": "^2.0.6",
-    "prettier": "2.2.1",
-    "typescript": "^4.2.3"
   },
   "keywords": [
     "gatsby"
@@ -51,18 +57,11 @@
   "scripts": {
     "build": "gatsby build",
     "develop": "gatsby develop",
-    "format": "prettier --write \"**/*.{js,jsx,ts,tsx,json,md}\"",
     "start": "npm run develop",
-    "serve": "gatsby serve",
-    "clean": "gatsby clean",
-    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1",
-    "type-check": "tsc --noEmit"
+    "format": "prettier --write \"**/*.+(js|jsx|tsx)\"",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/gatsbyjs/gatsby-starter-default"
-  },
-  "bugs": {
-    "url": "https://github.com/gatsbyjs/gatsby/issues"
+  "devDependencies": {
+    "@types/react-helmet": "^6.1.0"
   }
 }
