@@ -1,25 +1,22 @@
-const slugify = require("slugify")
-import IconList from 'src/components/IconList'
+import React from "react";
+import { IconList } from "./IconList";
+import { DistributionList } from "../hooks/DistributionList";
 
-export default function DownloadList(props: {
-  items: any[]
-}) {
-  let items = []
-  if (props.items[0].node !== undefined) {
-    props.items.map((item: { node: any }, index: any) => items.push(item.node))
-  } else {
-    items = props.items
-  }
-  items.forEach(function(item, index) {
-    this[index]['slug'] = slugify(item.name, { lower: true });
+const slugify = require("slugify");
+
+function DownloadList(props: { items: any[] }) {
+  const items = DistributionList();
+  items.forEach(function (item, index) {
+    this[index]["slug"] = slugify(item.name, { lower: true });
   }, items);
 
   return (
-      <IconList items={items}
-        linkroot='/download/'
-        width='90%'
-        marginleft='5%'
-        marginright='5%'
-      />
-  )
+    <IconList
+      items={items}
+      linkroot="/download/"
+      className="grid grid-cols-2 pt-4 pb-6 gap-4 lg:grid-cols-4"
+    />
+  );
 }
+
+export { DownloadList };
