@@ -3,14 +3,14 @@
 // format for console log when inspecting json
 // console.log(inspect(authors, { colors: true, depth: Infinity }));
 
-const fetch = require("node-fetch")
-const parse = require("xml-parser")
-const slugify = require("slugify")
-const fs = require("fs")
-const getargs = require('minimist')
-const inspect = require("util").inspect // this is only here to inspect the json during debugging
+const fetch = require("node-fetch");
+const parse = require("xml-parser");
+const slugify = require("slugify");
+const fs = require("fs");
+const getargs = require("minimist");
+const inspect = require("util").inspect; // this is only here to inspect the json during debugging
 
-const TODAY = new Date()
+const TODAY = new Date();
 const CATEGORIES = [
   { id: "kodi.audiodecoder", desc: "Audio decoders" },
   { id: "kodi.audioencoder", desc: "Audio encoders" },
@@ -74,111 +74,109 @@ const CATEGORIES = [
   { id: "xbmc.player.musicviz", desc: "Visualisations" },
   { id: "xbmc.python.weather", desc: "Weather providers" },
   { id: "xbmc.webinterface", desc: "Web interfaces" },
-]
+];
 /** @type {any[]} */
-let history = []
+let history = [];
 /** @type {any[]} */
-let kodiversion = ""
-let kodimirror = "http://ftp.halifax.rwth-aachen.de/xbmc/addons/"
-let kodistats = "http://mirrors.kodi.tv/.stats/redis_file_stats.json"
+let kodiversion = "";
+let kodimirror = "http://ftp.halifax.rwth-aachen.de/xbmc/addons/";
+let kodistats = "http://mirrors.kodi.tv/.stats/redis_file_stats.json";
 /** @type {import("../../src/addon").IAddon} */
-let addon = {}
+let addon = {};
 /** @type {import("../../src/addon").IAddon[]} */
-let addons = []
+let addons = [];
 // let featured = { addons: [] }
 /** @type {any[]} */
-let authors = []
+let authors = [];
 /** @type {any[]} */
-let categories = []
-let addonpath = ""
-let addonplatform = ""
-let addonstats = ""
+let categories = [];
+let addonpath = "";
+let addonplatform = "";
+let addonstats = "";
 /** @type {string[]} */
-let addonimagetypes = []
-let downloadqueue = []
-let versiondownloads = 0
-let gatsbyroot = "../../"
-let pixiememory = gatsbyroot + "src/data/addons/"
-let addonnodetype = ""
-let authornodetype = ""
-let categorynodetype = ""
-let data = ""
+let addonimagetypes = [];
+let downloadqueue = [];
+let versiondownloads = 0;
+let gatsbyroot = "../../";
+let pixiememory = gatsbyroot + "src/data/addons/";
+let addonnodetype = "";
+let authornodetype = "";
+let categorynodetype = "";
+let data = "";
 
 async function loadHistoryFile() {
-  return new Promise(function(resolve, reject) {
-    console.log('loading addons.json from ' + pixiememory)
+  return new Promise(function (resolve, reject) {
+    console.log("loading addons.json from " + pixiememory);
     try {
-      ah = JSON.parse(fs.readFileSync(pixiememory + 'addons.json', "utf8"))
+      ah = JSON.parse(fs.readFileSync(pixiememory + "addons.json", "utf8"));
     } catch (e) {
-      console.log("unable to load history, starting from scratch")
-      ah = [{}]
+      console.log("unable to load history, starting from scratch");
+      ah = [{}];
     }
-    resolve(ah)
-  })
+    resolve(ah);
+  });
 }
 
 function getAddon(rawaddon) {
-  let firstever = false
-  addon = addons.find(o => o.id === rawaddon.attributes.id)
-  addonpath = ""
-  addonplatform = ""
-  addonimagetypes = []
+  let firstever = false;
+  addon = addons.find(o => o.id === rawaddon.attributes.id);
+  addonpath = "";
+  addonplatform = "";
+  addonimagetypes = [];
   if (addon == undefined) {
-    addon = {}
-    addon.id = rawaddon.attributes.id
-    addon.addonid = addon.id
-    addon.name = rawaddon.attributes.name
-    addon.slug = slugify(rawaddon.attributes.id, { lower: true })
-    addon.version = rawaddon.attributes.version
-    addon.authors = []
-    addon.platforms = []
-    addon.categories = []
-    if (rawaddon.attributes["provider-name"] != undefined){
+    addon = {};
+    addon.id = rawaddon.attributes.id;
+    addon.addonid = addon.id;
+    addon.name = rawaddon.attributes.name;
+    addon.slug = slugify(rawaddon.attributes.id, { lower: true });
+    addon.version = rawaddon.attributes.version;
+    addon.authors = [];
+    addon.platforms = [];
+    addon.categories = [];
+    if (rawaddon.attributes["provider-name"] != undefined) {
       rawaddon.attributes["provider-name"]
         .split(",")
         .map(item => item.trim())
-        .forEach(assignAuthor)
+        .forEach(assignAuthor);
     }
-    addonhistory = history.find(o => o.id === addon.id)
+    addonhistory = history.find(o => o.id === addon.id);
     if (addonhistory == undefined) {
-      addonhistory = {}
-      addonhistory.id = addon.id
-      addonhistory.version = "none"
-      addonhistory.lastupdate = TODAY
-      addonhistory.firstseen = TODAY
-      addonhistory.agetype = "new"
-      addonhistory.downloads = 0
-      firstever = true
+      addonhistory = {};
+      addonhistory.id = addon.id;
+      addonhistory.version = "none";
+      addonhistory.lastupdate = TODAY;
+      addonhistory.firstseen = TODAY;
+      addonhistory.agetype = "new";
+      addonhistory.downloads = 0;
+      firstever = true;
     }
-    addon.firstseen = addonhistory.firstseen
-    addon.agetype = addonhistory.agetype
-    rawaddon.children.forEach(parseExtensions)
+    addon.firstseen = addonhistory.firstseen;
+    addon.agetype = addonhistory.agetype;
+    rawaddon.children.forEach(parseExtensions);
     if (addon.version == addonhistory.version) {
-      addon.lastupdate = addonhistory.lastupdate
-      addon.downloads = addonhistory.downloads
+      addon.lastupdate = addonhistory.lastupdate;
+      addon.downloads = addonhistory.downloads;
     } else {
-      addon.lastupdate = TODAY
+      addon.lastupdate = TODAY;
       if (!firstever) {
-        addon.agetype = "existing"
+        addon.agetype = "existing";
       }
-      addon.downloads = 0
+      addon.downloads = 0;
       if (addon.broken == null) {
-        queueImages()
+        queueImages();
       }
     }
     if (addon.broken == null) {
       if (addon.icons == null) {
-        addon.icons = [
-          { remotepath: "", localpath: "/images/default-addon.png" },
-        ]
+        addon.icons = [{ remotepath: "", localpath: "/images/default-addon.png" }];
       }
-      addon.icon = addon.icons[0].localpath
-      addons.push(addon)
-      addon.authors.forEach(createAuthorNode)
-      addon.categories.forEach(createCategoryNode)
+      addon.icon = addon.icons[0].localpath;
+      addons.push(addon);
+      addon.authors.forEach(createAuthorNode);
+      addon.categories.forEach(createCategoryNode);
     }
   } else {
-    rawaddon.children.forEach(parseExtensions)
+    rawaddon.children.forEach(parseExtensions);
   }
 }
 
@@ -186,41 +184,41 @@ function parseExtensions(extension) {
   if (extension.name == "extension") {
     switch (extension.attributes.point) {
       case "kodi.addon.metadata":
-        extension.children.forEach(getMetadata)
+        extension.children.forEach(getMetadata);
         addon.platforms.push({
           platform: addonplatform,
           path: addonpath,
           statspath: addonstatspath,
-        })
-        break
+        });
+        break;
       case "xbmc.addon.metadata":
-        extension.children.forEach(getMetadata)
+        extension.children.forEach(getMetadata);
         addon.platforms.push({
           platform: addonplatform,
           path: addonpath,
           statspath: addonstatspath,
-        })
-        break
+        });
+        break;
       case "kodi.python.pluginsource":
-        assignCategory("Plugins")
-        checkProvides(extension.children)
-        break
+        assignCategory("Plugins");
+        checkProvides(extension.children);
+        break;
       case "xbmc.python.pluginsource":
-        assignCategory("Plugins")
-        checkProvides(extension.children)
-        break
+        assignCategory("Plugins");
+        checkProvides(extension.children);
+        break;
       case "kodi.python.script":
-        assignCategory("Program addons")
-        checkProvides(extension.children)
-        break
+        assignCategory("Program addons");
+        checkProvides(extension.children);
+        break;
       case "xbmc.python.script":
-        assignCategory("Program addons")
-        checkProvides(extension.children)
-        break
+        assignCategory("Program addons");
+        checkProvides(extension.children);
+        break;
       default:
-        cat = CATEGORIES.find(o => o.id === extension.attributes.point)
+        cat = CATEGORIES.find(o => o.id === extension.attributes.point);
         if (cat != undefined) {
-          assignCategory(cat.desc)
+          assignCategory(cat.desc);
         }
     }
   }
@@ -228,51 +226,51 @@ function parseExtensions(extension) {
 
 function checkProvides(provider) {
   if (provider.length > 0) {
-    if (provider[0].content == undefined){
-      return
+    if (provider[0].content == undefined) {
+      return;
     }
     if (provider[0].content.includes("audio")) {
-      assignCategory("Music addons")
+      assignCategory("Music addons");
     }
     if (provider[0].content.includes("video")) {
-      assignCategory("Video addons")
+      assignCategory("Video addons");
     }
     if (provider[0].content.includes("image")) {
-      assignCategory("Picture addons")
+      assignCategory("Picture addons");
     }
     if (provider[0].content.includes("executable")) {
-      assignCategory("Program addons")
+      assignCategory("Program addons");
     }
   }
 }
 
 function createAuthorNode(author) {
-  newauthor = JSON.parse(JSON.stringify(author))
-  authorcheck = authors.find(o => o.id === newauthor.name)
+  newauthor = JSON.parse(JSON.stringify(author));
+  authorcheck = authors.find(o => o.id === newauthor.name);
   if (authorcheck == undefined) {
-    newauthor.id = newauthor.name
-    newauthor.addons = [addon]
-    authors.push(newauthor)
+    newauthor.id = newauthor.name;
+    newauthor.addons = [addon];
+    authors.push(newauthor);
   } else {
-    addoncheck = authorcheck.addons.find(o => o.id === addon.id)
+    addoncheck = authorcheck.addons.find(o => o.id === addon.id);
     if (addoncheck == undefined) {
-      authorcheck.addons.push(addon)
+      authorcheck.addons.push(addon);
     }
   }
 }
 
 function createCategoryNode(category) {
-  newcategory = JSON.parse(JSON.stringify(category))
-  categorycheck = categories.find(o => o.id === newcategory.name)
+  newcategory = JSON.parse(JSON.stringify(category));
+  categorycheck = categories.find(o => o.id === newcategory.name);
   if (categorycheck == undefined) {
-    newcategory.id = newcategory.name
-    newcategory.grouping = setCategoryGrouping(newcategory.name)
-    newcategory.addons = [addon]
-    categories.push(newcategory)
+    newcategory.id = newcategory.name;
+    newcategory.grouping = setCategoryGrouping(newcategory.name);
+    newcategory.addons = [addon];
+    categories.push(newcategory);
   } else {
-    addoncheck = categorycheck.addons.find(o => o.id === addon.id)
+    addoncheck = categorycheck.addons.find(o => o.id === addon.id);
     if (addoncheck == undefined) {
-      categorycheck.addons.push(addon)
+      categorycheck.addons.push(addon);
     }
   }
 }
@@ -280,203 +278,212 @@ function createCategoryNode(category) {
 function setCategoryGrouping(name) {
   switch (name.toLowerCase()) {
     case "game clients":
-      return "Games"
+      return "Games";
     case "controller profiles":
-      return "Games"
+      return "Games";
     case "album information providers":
-      return "Information providers"
+      return "Information providers";
     case "artist information providers":
-      return "Information providers"
+      return "Information providers";
     case "movie information providers":
-      return "Information providers"
+      return "Information providers";
     case "music video information providers":
-      return "Information providers"
+      return "Information providers";
     case "tv information providers":
-      return "Information providers"
+      return "Information providers";
     case "image collections":
-      return "Look and feel"
+      return "Look and feel";
     case "languages":
-      return "Look and feel"
+      return "Look and feel";
     case "screensavers":
-      return "Look and feel"
+      return "Look and feel";
     case "skins":
-      return "Look and feel"
+      return "Look and feel";
     case "gui sounds":
-      return "Look and feel"
+      return "Look and feel";
     default:
-      return "Other"
+      return "Other";
   }
 }
 
 function assignAuthor(author) {
-  authorcheck = addon.authors.find(o => o.name === author)
+  authorcheck = addon.authors.find(o => o.name === author);
   if (authorcheck == undefined) {
-    slug = slugify(author, { lower: true })
-    icon = "/images/authors/" + slug + ".png"
-    addon.authors.push({ name: author, slug: slug, icon: icon })
+    slug = slugify(author, { lower: true });
+    icon = "/images/authors/" + slug + ".png";
+    addon.authors.push({ name: author, slug: slug, icon: icon });
   }
 }
 
 function assignCategory(category) {
-  categorycheck = addon.categories.find(o => o.name === category)
+  categorycheck = addon.categories.find(o => o.name === category);
   if (categorycheck == undefined) {
-    slug = slugify(category, { lower: true })
-    icon = "/images/categories/" + slug + ".png"
-    addon.categories.push({ name: category, slug: slug, icon: icon })
+    slug = slugify(category, { lower: true });
+    icon = "/images/categories/" + slug + ".png";
+    addon.categories.push({ name: category, slug: slug, icon: icon });
   }
 }
 
 function getMetadata(metadata) {
   if (metadata.name == "summary" || metadata.name == "description") {
     if (metadata.attributes.lang == "en_GB") {
-      addon[metadata.name] = metadata.content
+      addon[metadata.name] = metadata.content;
     }
   } else if (metadata.name == "size") {
     addon[metadata.name] =
-      String(Math.floor(Number(metadata.content) / 1024)) + "KB"
+      String(Math.floor(Number(metadata.content) / 1024)) + "KB";
   } else if (metadata.name == "path") {
-    addonpath = kodimirror + metadata.content
-    addonstatspath = '/addons/' + kodiversion + '/' + metadata.content
+    addonpath = kodimirror + metadata.content;
+    addonstatspath = "/addons/" + kodiversion + "/" + metadata.content;
   } else if (metadata.name == "platform") {
-    addonplatform = metadata.content
+    addonplatform = metadata.content;
   } else if (metadata.name == "assets") {
-    metadata.children.forEach(getAssets)
+    metadata.children.forEach(getAssets);
   } else {
-    addon[metadata.name] = metadata.content
+    addon[metadata.name] = metadata.content;
   }
 }
 
 function getAssets(asset) {
-  let assetcheck = undefined
-  let arrayname = asset.name + "s"
+  let assetcheck = undefined;
+  let arrayname = asset.name + "s";
   if (addon[arrayname] == undefined) {
-    addon[arrayname] = []
-    addonimagetypes.push(arrayname)
+    addon[arrayname] = [];
+    addonimagetypes.push(arrayname);
   } else {
-    assetcheck = addon[arrayname].find(o => o.remotepath === asset.content)
+    assetcheck = addon[arrayname].find(o => o.remotepath === asset.content);
   }
   if (assetcheck == undefined) {
     imagepath =
-      "/images/addons/" + kodiversion + "/" +
+      "/images/addons/" +
+      kodiversion +
+      "/" +
       slugify(addon.id, { lower: true }) +
       "/" +
-      asset.content
-    addon[arrayname].push({ localpath: imagepath, remotepath: asset.content })
+      asset.content;
+    addon[arrayname].push({ localpath: imagepath, remotepath: asset.content });
   }
 }
 
 function doCleanup(item) {
-  item.addons.sort(compare)
-  item.totaladdons = item.addons.length
+  item.addons.sort(compare);
+  item.totaladdons = item.addons.length;
 }
 
 function compare(a, b) {
-  const bandA = a.name.toUpperCase()
-  const bandB = b.name.toUpperCase()
-  let comparison = 0
+  const bandA = a.name.toUpperCase();
+  const bandB = b.name.toUpperCase();
+  let comparison = 0;
   if (bandA > bandB) {
-    comparison = 1
+    comparison = 1;
   } else if (bandA < bandB) {
-    comparison = -1
+    comparison = -1;
   }
-  return comparison
+  return comparison;
 }
 
 function queueImages() {
-  addonimagetypes.forEach(queueImageType)
+  addonimagetypes.forEach(queueImageType);
 }
 
 function queueImageType(imagetype) {
-  const fullurl = addon.platforms[0].path
-  const urlbase = fullurl.substring(0, fullurl.lastIndexOf("/")) + "/"
-  const rootpath = gatsbyroot + "static"
+  const fullurl = addon.platforms[0].path;
+  const urlbase = fullurl.substring(0, fullurl.lastIndexOf("/")) + "/";
+  const rootpath = gatsbyroot + "static";
   addon[imagetype].forEach(asset => {
-    const localpath = asset.localpath
+    const localpath = asset.localpath;
     const localbase =
-      rootpath + localpath.substring(0, localpath.lastIndexOf("/")) + "/"
-    downloadqueue.push({remote: urlbase + asset.remotepath, localdir: localbase, local: rootpath + asset.localpath })
-  })
+      rootpath + localpath.substring(0, localpath.lastIndexOf("/")) + "/";
+    downloadqueue.push({
+      remote: urlbase + asset.remotepath,
+      localdir: localbase,
+      local: rootpath + asset.localpath,
+    });
+  });
 }
 
 function getDownloadCount() {
   if (addon.downloads === undefined) {
-    addon.downloads = 0
+    addon.downloads = 0;
   }
 }
 
 async function app() {
-  const args = getargs(process.argv.slice(2))
-  if (args['kv'] == undefined) {
-    console.log('kodi version not included on command line')
-    console.log('use --kv=<version>')
-    return
+  const args = getargs(process.argv.slice(2));
+  if (args["kv"] == undefined) {
+    console.log("kodi version not included on command line");
+    console.log("use --kv=<version>");
+    return;
   }
-  console.log('kodi version from command line is ' + args['kv'])
-  kodiversion = args['kv']
-  kodimirror = kodimirror  + kodiversion + "/"
-  pixiememory = pixiememory + kodiversion + '/'
-  history = await loadHistoryFile()
-  console.log('getting addons from the ' + kodiversion + ' repo using ' + kodimirror)
+  console.log("kodi version from command line is " + args["kv"]);
+  kodiversion = args["kv"];
+  kodimirror = kodimirror + kodiversion + "/";
+  pixiememory = pixiememory + kodiversion + "/";
+  history = await loadHistoryFile();
+  console.log(
+    "getting addons from the " + kodiversion + " repo using " + kodimirror
+  );
   try {
-    const res = await fetch(kodimirror + 'addons.xml')
-    data = await res.text()
+    const res = await fetch(kodimirror + "addons.xml");
+    data = await res.text();
   } catch (error) {
-    data = ''
-    console.log(error)
+    data = "";
+    console.log(error);
   }
-  console.log('getting addon download stats from ' + kodistats)
+  console.log("getting addon download stats from " + kodistats);
   try {
-    const res = await fetch(kodistats)
-    rawstats = await res.json()
-    let firstKey = Object.keys(rawstats[0])[0]
-    addonstats = rawstats[0][firstKey]
+    const res = await fetch(kodistats);
+    rawstats = await res.json();
+    let firstKey = Object.keys(rawstats[0])[0];
+    addonstats = rawstats[0][firstKey];
   } catch (error) {
-    addonstats = {}
-    console.log(error)
+    addonstats = {};
+    console.log(error);
   }
   if (data) {
-    const parsedXML = parse(data)
-    console.log('parsing addon data')
-    parsedXML.root.children.forEach(getAddon)
-    authors.forEach(doCleanup)
-    categories.forEach(doCleanup)
-    for (let i=0; i < downloadqueue.length; i++) {
-      let createddir = false
-      let download = downloadqueue[i]
+    const parsedXML = parse(data);
+    console.log("parsing addon data");
+    parsedXML.root.children.forEach(getAddon);
+    authors.forEach(doCleanup);
+    categories.forEach(doCleanup);
+    for (let i = 0; i < downloadqueue.length; i++) {
+      let createddir = false;
+      let download = downloadqueue[i];
       await fs.mkdir(download.localdir, { recursive: true }, err => {
         if (err) {
-          console.log('create failed for directory' + download.localdir)
-        } 
-      })
-      console.log('downloading ' + download.remote)
+          console.log("create failed for directory" + download.localdir);
+        }
+      });
+      console.log("downloading " + download.remote);
       await fetch(download.remote).then(res => {
-        const dest = fs.createWriteStream(download.local)
-        res.body.pipe(dest)
-      })
+        const dest = fs.createWriteStream(download.local);
+        res.body.pipe(dest);
+      });
     }
-    console.log('getting addon download counts')
-    for (let i=0; i < addons.length; i++) {
-      let downloadcount = 0
-      for (let k=0; k < addons[i].platforms.length; k++) {
-        let pcstring = addonstats[addons[i].platforms[k].statspath]
+    console.log("getting addon download counts");
+    for (let i = 0; i < addons.length; i++) {
+      let downloadcount = 0;
+      for (let k = 0; k < addons[i].platforms.length; k++) {
+        let pcstring = addonstats[addons[i].platforms[k].statspath];
         if (pcstring != undefined) {
-          let platformcount = parseInt(pcstring)
-          downloadcount = downloadcount + platformcount
+          let platformcount = parseInt(pcstring);
+          downloadcount = downloadcount + platformcount;
         }
       }
-      if (downloadcount < addons[i].downloads) { // if true a new year has started, which resets the stats
-        addons[i].downloads = addons[i].downloads + downloadcount
+      if (downloadcount < addons[i].downloads) {
+        // if true a new year has started, which resets the stats
+        addons[i].downloads = addons[i].downloads + downloadcount;
       } else {
-        addons[i].downloads = downloadcount
+        addons[i].downloads = downloadcount;
       }
     }
-    console.log('writing addons.json to ' + pixiememory)
-    fs.writeFileSync(pixiememory + 'addons.json', JSON.stringify(addons))
-    console.log('writing authors.json to ' + pixiememory)
-    fs.writeFileSync(pixiememory + 'authors.json', JSON.stringify(authors))
-    console.log('writing categories.json to ' + pixiememory)
-    fs.writeFileSync(pixiememory + 'categories.json', JSON.stringify(categories))
+    console.log("writing addons.json to " + pixiememory);
+    fs.writeFileSync(pixiememory + "addons.json", JSON.stringify(addons));
+    console.log("writing authors.json to " + pixiememory);
+    fs.writeFileSync(pixiememory + "authors.json", JSON.stringify(authors));
+    console.log("writing categories.json to " + pixiememory);
+    fs.writeFileSync(pixiememory + "categories.json", JSON.stringify(categories));
   }
 }
 
-app()
+app();
