@@ -1,180 +1,226 @@
-import React from "react"
-import {
-  Text,
-  TextVariants,
-  TextContent,
-  TextList,
-  TextListItem,
-  List,
-  ListItem,
-  ListVariant,
-  Split,
-  SplitItem,
-  Stack,
-  StackItem,
-} from "@patternfly/react-core"
-import WarningTriangleIcon from "@patternfly/react-icons/dist/js/icons/warning-triangle-icon"
-import ConditionalSlideshow from "src/components/ConditionalSlideshow"
-import ConditionalTextListItem from "src/components/ConditionalTextListItem"
-import ItemWithComma from "src/components/ItemWithComma"
-import Layout from "src/gatsby-theme-patternfly/components/Layout"
-import { IAddon } from "src/addon"
-import MetadataHeader from "src/components/SiteMetadata"
+import React from "react";
+import ReactMarkdown from "react-markdown";
+import { ExclamationOutline } from "heroicons-react";
+import { IAddon } from "./../addon";
+import { DefaultLayout } from "./Layout";
+import ItemWithComma from "./ItemWithComma";
+import { Card } from "./Card";
+import { Carousel } from "./Carousel";
 
 export default function AddonLayout({
+  repo,
   data,
 }: {
+  repo: string;
   data: {
-    addon: IAddon
-  }
+    addon: IAddon;
+  };
 }) {
-  const addon = data.addon
-  let slides: { source: string }[] = []
-  let authorlabel = "Author"
-  let categorylabel = "Category"
+  const addon = data.addon;
+  let slides: string[] = [];
+  let authorlabel = "Author";
+  let categorylabel = "Category";
   if (addon.authors.length > 1) {
-    authorlabel = "Authors"
+    authorlabel = "Authors";
   }
   if (addon.categories.length > 1) {
-    categorylabel = "Categories"
+    categorylabel = "Categories";
   }
   if (addon.screenshots != null) {
     addon.screenshots.forEach((screenshot: { localpath: string }) => {
-      slides.push({ source: screenshot.localpath })
-    })
+      slides.push(screenshot.localpath);
+    });
   }
+  let frontmatter = {
+    title: repo + " Add-on Details",
+    breadcrumbs: "Addons | " + repo + " | " + addon.name,
+  };
 
   return (
-    <Layout>
-      <MetadataHeader title={addon.name + ' | Add-On'} />
-      <Stack hasGutter>
-        <StackItem>
-          <Split hasGutter>
-            <SplitItem>
-              <img width="150" height="150" alt="" src={addon.icon} />
-            </SplitItem>
-            <SplitItem isFilled>
-              <Stack hasGutter>
-                <StackItem>
-                  <TextContent>
-                    <Text component={TextVariants.h1}><div dangerouslySetInnerHTML={{ __html: addon.name }} /></Text>
-                    <Text><div dangerouslySetInnerHTML={{ __html: addon.summary }} /></Text>
-                  </TextContent>
-                </StackItem>
-                <StackItem>
-                  <List variant={ListVariant.inline}>
-                    <ListItem>&#128194; {addon.version}</ListItem>
-                    <ListItem>&#128339; {addon.lastupdate}</ListItem>
-                    <ListItem>&#128100;&nbsp;
-                      {addon.authors.map(
-                        (
-                          author: { name: string; slug: string },
-                          index: any
-                        ) => {
-                          return (
-                            <ItemWithComma
-                              description={author.name}
-                              url={"../author/" + author.slug}
-                              index={index}
-                              length={addon.authors.length - 1}
-                              linkType="internal"
-                            />
-                          )
-                        }
-                      )}
-                    </ListItem>
-                  </List>
-                  <List variant={ListVariant.inline}>
-                    <ListItem>&#127991;&nbsp;
-                      {addon.categories.map(
-                        (category: { name: string; slug: string }, index: any) => {
-                          return (
-                            <ItemWithComma
-                              description={category.name}
-                              url={"../category/" + category.slug}
-                              index={index}
-                              length={addon.categories.length - 1}
-                              linkType="internal"
-                            />
-                          )
-                        }
-                      )}
-                    </ListItem>
-                  </List>
-                </StackItem>
-              </Stack>
-            </SplitItem>
-          </Split>
-        </StackItem>
-        <StackItem>
-          <Split hasGutter>
-            <SplitItem>
-              <WarningTriangleIcon />
-            </SplitItem>
-            <SplitItem isFilled>
-              To download this Add-On, we highly recommend you do it via the
-              user interface in Kodi. Simply look for the "Get More" button in
-              the Add-Ons menu. If you want to install it manually, you can
-              direct download from the platforms link that matches your platform
-              then in Kodi look for the "Install via Zip" option.
-            </SplitItem>
-          </Split>
-        </StackItem>
-        <StackItem>
-          <TextContent id="info" className="pf-u-py-xl">
-            <TextList component="dl">
-              <TextListItem component="dt">Description</TextListItem>
-              <TextListItem component="dd"><div dangerouslySetInnerHTML={{ __html: addon.description }} /></TextListItem>
-              <ConditionalTextListItem
-                hide={addon.forum == null}
-                title={"Forum"}
-                description={addon.forum}
-                url={addon.forum}
-              />
-              <ConditionalTextListItem
-                hide={addon.website == null}
-                title={"Website"}
-                description={addon.website}
-                url={addon.website}
-              />
-              <ConditionalTextListItem
-                hide={addon.source == null}
-                title={"Source"}
-                description={addon.source}
-                url={addon.source}
-              />
-              <TextListItem component="dt">License</TextListItem>
-              <TextListItem component="dd">{addon.license}</TextListItem>
-              <TextListItem component="dt">Platforms</TextListItem>
-              <TextListItem component="dd">
-                {addon.platforms.map(
-                  (
-                    platform: { platform: string; path: string },
-                    index: any
-                  ) => {
+    <DefaultLayout frontmatter={frontmatter}>
+      <div className="flex flex-col">
+        <div className="grid grid-cols-5">
+          <div align="right" className="col-span-1 row-span-full pr-6">
+            <img width="150" height="150" alt="" src={addon.icon} />
+          </div>
+          <div className="col-span-4 flex flex-col">
+            <ReactMarkdown className="text-2xl font-bold">
+              {addon.name}
+            </ReactMarkdown>
+            <p>
+              <ReactMarkdown>{addon.summary}</ReactMarkdown>
+            </p>
+            <div className="flex prose prose-blue">
+              <div className="pr-4">&#128194; {addon.version}</div>
+              <div className="pr-4">&#128339; {addon.lastupdate}</div>
+              <div>
+                &#128100;&nbsp;
+                {addon.authors.map(
+                  (author: { name: string; slug: string }, index: any) => {
                     return (
                       <ItemWithComma
-                        description={platform.platform}
-                        url={platform.path}
+                        description={author.name}
+                        url={"../author/" + author.slug}
                         index={index}
-                        length={addon.platforms.length - 1}
-                        linkType="external"
+                        length={addon.authors.length - 1}
+                        linkType="internal"
                       />
-                    )
+                    );
                   }
                 )}
-              </TextListItem>
-              <TextListItem component="dt">Size</TextListItem>
-              <TextListItem component="dd">{addon.size}</TextListItem>
-              <TextListItem component="dt">Downloads</TextListItem>
-              <TextListItem component="dd">{addon.downloads.toLocaleString()} (this is the download count for the most current version)</TextListItem>
-            </TextList>
-          </TextContent>
-        </StackItem>
-        <StackItem>
-          <ConditionalSlideshow slides={slides} />
-        </StackItem>
-      </Stack>
-    </Layout>
-  )
+              </div>
+            </div>
+            <div className="flex prose prose-blue">
+              <div>
+                &#127991;&nbsp;
+                {addon.categories.map(
+                  (category: { name: string; slug: string }, index: any) => {
+                    return (
+                      <ItemWithComma
+                        description={category.name}
+                        url={"../category/" + category.slug}
+                        index={index}
+                        length={addon.categories.length - 1}
+                        linkType="internal"
+                      />
+                    );
+                  }
+                )}
+              </div>
+            </div>
+          </div>
+        </div>
+        <div className="pt-2 px-2 md:pt-4 md:px-10 lg:px-32">
+          <Card>
+            <div className="flex p-2">
+              <div className="pr-2">
+                <ExclamationOutline className="inline text-kodi" />
+              </div>
+              <div>
+                To download this Add-On, we highly recommend you do it via the user
+                interface in Kodi. Simply look for the "Get More" button in the
+                Add-Ons menu. If you want to install it manually, you can direct
+                download from the platforms link that matches your platform then in
+                Kodi look for the "Install via Zip" option.
+              </div>
+            </div>
+          </Card>
+        </div>
+        <div className="grid grid-rows-none md:grid-cols-5 pt-4 gap-4 prose prose-blue max-w-none">
+          <div
+            className="col-span-1 font-bold md:text-right pr-2"
+            style={{ display: addon.description == null ? "none" : "block" }}
+          >
+            {"Description"}
+          </div>
+          <div
+            className="md:col-span-4"
+            style={{ display: addon.description == null ? "none" : "block" }}
+          >
+            <ReactMarkdown>{addon.description}</ReactMarkdown>
+          </div>
+          <div
+            className="pr-2 font-bold md:col-span-1 md:text-right"
+            style={{ display: addon.forum == null ? "none" : "block" }}
+          >
+            {"Forum"}
+          </div>
+          <div
+            className="md:col-span-4"
+            style={{ display: addon.forum == null ? "none" : "block" }}
+          >
+            <a href={addon.forum}>{addon.forum}</a>
+          </div>
+          <div
+            className="pr-2 font-bold md:col-span-1 md:text-right"
+            style={{ display: addon.website == null ? "none" : "block" }}
+          >
+            {"Website"}
+          </div>
+          <div
+            className="md:col-span-4"
+            style={{ display: addon.website == null ? "none" : "block" }}
+          >
+            <a href={addon.website}>{addon.website}</a>
+          </div>
+          <div
+            className="pr-2 font-bold md:col-span-1 md:text-right"
+            style={{ display: addon.source == null ? "none" : "block" }}
+          >
+            {"Source"}
+          </div>
+          <div
+            className="md:col-span-4"
+            style={{ display: addon.source == null ? "none" : "block" }}
+          >
+            <a href={addon.source}>{addon.source}</a>
+          </div>
+          <div
+            className="pr-2 font-bold md:col-span-1 md:text-right"
+            style={{ display: addon.license == null ? "none" : "block" }}
+          >
+            {"License"}
+          </div>
+          <div
+            className="md:col-span-4"
+            style={{ display: addon.license == null ? "none" : "block" }}
+          >
+            {addon.license}
+          </div>
+          <div
+            className="pr-2 font-bold md:col-span-1 md:text-right"
+            style={{ display: addon.platforms == null ? "none" : "block" }}
+          >
+            {"Platforms"}
+          </div>
+          <div
+            className="md:col-span-4"
+            style={{ display: addon.platforms == null ? "none" : "block" }}
+          >
+            {addon.platforms.map(
+              (platform: { platform: string; path: string }, index: any) => {
+                return (
+                  <ItemWithComma
+                    description={platform.platform}
+                    url={platform.path}
+                    index={index}
+                    length={addon.platforms.length - 1}
+                    linkType="external"
+                  />
+                );
+              }
+            )}
+          </div>
+          <div
+            className="pr-2 font-bold md:col-span-1 md:text-right"
+            style={{ display: addon.size == null ? "none" : "block" }}
+          >
+            {"Size"}
+          </div>
+          <div
+            className="md:col-span-4"
+            style={{ display: addon.size == null ? "none" : "block" }}
+          >
+            {addon.size}
+          </div>
+          <div
+            className="pr-2 font-bold md:col-span-1 md:text-right"
+            style={{ display: addon.downloads == null ? "none" : "block" }}
+          >
+            {"Downloads"}
+          </div>
+          <div
+            className="md:col-span-4"
+            style={{ display: addon.downloads == null ? "none" : "block" }}
+          >
+            {addon.downloads.toLocaleString()} (this is the download count for the
+            most current version)
+          </div>
+        </div>
+        <div className="pt-6">
+          <Carousel slides={slides} />
+        </div>
+      </div>
+    </DefaultLayout>
+  );
 }
