@@ -1,60 +1,47 @@
-import React from 'react';
-import ReactMarkdown from 'react-markdown'
-import {
-  Tabs, Tab, TabTitleText,
-  Text, TextContent, TextVariants,
-  Stack, StackItem,
-} from '@patternfly/react-core'
-import { DownloadLinks } from './ReleasesContent'
+import React from "react";
+import ReactMarkdown from "react-markdown";
+import { DownloadLinks } from "./ReleasesLinks";
+import Tabs from "./Tabs";
 
 class ReleasesTabs extends React.Component {
-  constructor(private props: {releases: {name: string, id: string, downloads: {
-    url: string | undefined;
-    name: React.ReactNode;
-  }[], description: string}[]}) {
+  constructor(
+    private props: {
+      releases: {
+        name: string;
+        id: string;
+        downloads: {
+          url: string | undefined;
+          name: React.ReactNode;
+        }[];
+        description: string;
+      }[];
+    }
+  ) {
     super(props);
-    this.state = {
-      activeTabKey: 0,
-      isBox: true
-    }
-    // Toggle currently active tab
-    this.handleTabClick = (event: any, tabIndex: any) => {
-      this.setState({
-        activeTabKey: tabIndex,
-        isBox: true
-      })
-    }
   }
 
   render() {
-    const releases = this.props.releases
-    const {activeTabKey, isBox} = this.state;
+    const releases = this.props.releases;
     return (
-      <div>
-        <Tabs activeKey={activeTabKey} onSelect={this.handleTabClick} isBox={isBox}>
-          {releases.map((release, index) => {
-            return (
-              <Tab eventKey={index} title=<TabTitleText>{release.name}</TabTitleText>>
-                <Stack hasGutter>
-                  <StackItem>
-                    <div style={{'paddingTop': '15px'}}>
-                      <TextContent>
-                        <Text component={TextVariants.h2}>{release.title}</Text>
-                      </TextContent>
-                      <TextContent>
-                        <Text component={TextVariants.p}><ReactMarkdown>{release.description}</ReactMarkdown></Text>
-                      </TextContent>
-                    </div>
-                  </StackItem>
-                  <StackItem>
-                    <DownloadLinks downloads={release.downloads} />
-                  </StackItem>
-                </Stack>
-              </Tab>
-             )
-          })}
-        </Tabs>
-      </div>
+      <Tabs>
+        {releases.map((release, index) => {
+          return (
+            <div label={release.name} className="flex">
+              <div className="flex-initial">
+                <div className="pt-15">
+                  <h1 className="text-xl font-bold">{release.title}</h1>
+                  <ReactMarkdown className="prose prose-blue max-w-none">
+                    {release.description}
+                  </ReactMarkdown>
+                </div>
+              </div>
+              <div className="flex-initial pt-4">
+                <DownloadLinks downloads={release.downloads} />
+              </div>
+            </div>
+          );
+        })}
+      </Tabs>
     );
   }
 }
